Moving accordion methods into misc handler
