+++ module/sheets/actor-sheet.mjs
@@ -13,7 +13,7 @@ import { prepareZords, onZordDelete, onMorph } from "../sheet-handlers/power-ran
 import { onFocusDelete, onRoleDelete } from "../sheet-handlers/role-handler.mjs";
 import { onAltModeDelete, onTransform } from "../sheet-handlers/transformer-handler.mjs";
 import { onPerkDelete } from "../sheet-handlers/perk-handler.mjs";
-import { onRest, onRoll } from "../sheet-handlers/listener-misc-handler.mjs";
+import { onRest, onRoll, onToggleAccordion, onToggleHeaderAccordion } from "../sheet-handlers/listener-misc-handler.mjs";
 import { onDropActor, onDropItem } from "../sheet-handlers/drop-handler.mjs";
 
 export class Essence20ActorSheet extends ActorSheet {
@@ -363,10 +363,10 @@ export class Essence20ActorSheet extends ActorSheet {
     }
 
     // Open and collapse Item content
-    html.find('.accordion-label').click(this._onToggleAccordion.bind(this));
+    html.find('.accordion-label').click(ev => onToggleAccordion(ev, this));
 
     // Open and collapse all Item contents in container
-    html.find('.header-accordion-label').click(this._onToggleHeaderAccordion.bind(this));
+    html.find('.header-accordion-label').click(ev => onToggleHeaderAccordion(ev, this));
 
     // Drag events for macros.
     if (this.actor.isOwner) {
@@ -412,59 +412,6 @@ export class Essence20ActorSheet extends ActorSheet {
       html.find('.no-unlock').attr('readonly', true);
       html.find('select').attr('disabled', isLocked);
     });
-  }
-
-  /**
-   * Handle toggling accordion container headers.
-   * @param {Event} event The originating click event
-   * @private
-   */
-  async _onToggleHeaderAccordion(event) {
-    const el = event.currentTarget;
-    const isOpening = !$(el.closest('.header-accordion-wrapper')).hasClass('open');
-    $(el.closest('.header-accordion-wrapper')).toggleClass('open');
-
-    const accordionLabels = el.closest('.collapsible-item-container').querySelectorAll('.accordion-wrapper');
-    for (const accordionLabel of accordionLabels) {
-      isOpening ? $(accordionLabel).addClass('open') : $(accordionLabel).removeClass('open');
-    }
-  }
-
-  /**
-   * Handle toggling accordion containers.
-   * @param {Event} event The originating click event
-   * @private
-   */
-  async _onToggleAccordion(event) {
-    const el = event.currentTarget;
-    const parent = $(el).closest('.accordion-wrapper');
-
-    // Avoid collapsing NPC skills container on rerender
-    if (parent.hasClass('skills-container')) {
-      const isOpen = this._accordionStates.skills;
-      this._accordionStates.skills = isOpen ? '' : 'open';
-      this.render();
-    } else {
-      parent.toggleClass('open');
-
-      // Check if the container header toggle should be flipped
-      let oneClosed = false;
-
-      // Look for a closed Item
-      const accordionLabels = el.closest('.collapsible-item-container').querySelectorAll('.accordion-wrapper');
-      for (const accordionLabel of accordionLabels) {
-        oneClosed = !$(accordionLabel).hasClass('open');
-        if (oneClosed) break;
-      }
-
-      // Set header state to open if all Items are open; closed otherwise
-      const container = el.closest('.collapsible-item-container').querySelector('.header-accordion-wrapper');
-      if (oneClosed) {
-        $(container).removeClass('open');
-      } else {
-        $(container).addClass('open');
-      }
-    }
   }
 
   /**
